feat: extract log table and re-use in evidence viewer
