+++ src/components/Loading/index.js
@@ -1,10 +1,13 @@
 export function Loading({ text = "Loading..." }) {
   return (
-    <div class="flex justify-center items-center">
-      <div class="spinner-border animate-spin inline-block w-6 h-6 border-2 rounded-full" role="status">
-        <span class="visually-hidden"> </span>
+    <div className="flex justify-center items-center">
+      <div
+        className="spinner-border animate-spin inline-block w-6 h-6 border-2 rounded-full"
+        role="status"
+      >
+        <span className="visually-hidden"> </span>
       </div>
-      <span class="text-sm ml-3">{text}</span>
+      <span className="text-sm ml-3">{text}</span>
     </div>
   );
 }
